Makefile: Remove `deb` rule and use `find` to set permissions
The `deb` rule provided by the Makefile is not necessary for builds
using foundry and can be removed. Further, the Makefile does not use
`find` for setting permissions during installation, making changes
necessary when new modules are added.

This commit removes the `deb` rule from the Makefile and changes the
`install` rule to make use of `find` to set permissions on files and
directories.